Refactor: Extract class `Component`
