+++ Resources/Private/JavaScript/VideoPlayer/SimpleModal.js
@@ -1,11 +1,18 @@
 import $ from 'jquery';
 
-export default class SimpleModal {
+import Component from './Component';
+
+export default class SimpleModal extends Component {
   /**
    *
    * @param {HTMLElement} parent
    */
   constructor(parent, state = {}) {
+    super({
+      show: false,
+      ...state,
+    });
+
     this._parent = parent;
 
     /**
@@ -14,15 +21,6 @@ export default class SimpleModal {
      * modal visibility.
      */
     this._isAnimating = false;
-
-    this._state = {
-      show: false,
-      ...state,
-    };
-
-    this._pendingStateUpdates = [];
-    this._renderTimeout = null;
-    this._renderPromise = Promise.resolve();
 
     this._dom = this._createDom();
     this._dom.close.addEventListener('click', this.close.bind(this));
@@ -75,31 +73,6 @@ export default class SimpleModal {
 
   toggle() {
     this.open(!this._state.show);
-  }
-
-  update() {
-    return this._renderPromise;
-  }
-
-  setState(state = {}) {
-    this._pendingStateUpdates.push(state);
-
-    // Postpone updates so that multiple synchronous calls to `setState` don't
-    // lead to multiple renderings.
-    if (!this._renderTimeout) {
-      this._renderPromise = new Promise(resolve => {
-        this._renderTimeout = setTimeout(() => {
-          const state = this._pendingStateUpdates.reduce(Object.assign, {});
-          const newState = Object.assign({}, this._state, state);
-          this.render(newState);
-          this._state = newState;
-          this._pendingStateUpdates = [];
-          this._renderTimeout = null;
-          this._renderPromise = Promise.resolve();
-          resolve();
-        });
-      })
-    }
   }
 
   render(state) {
